+++ src/Controller/KatalogPageAddController.java
@@ -253,7 +253,7 @@ public class KatalogPageAddController implements Initializable {
 
         }else{
         try {
-            DB.InsertBarang(NamaBarang,HargaBarang,PenjualBarang,KondisiBarang,UkuranBarang,BrandBarang,WarnaBarang,KategoriBarang,Deskripsi,GambarBarang);
+            DB.InsertBarang(NamaBarang,HargaBarang,PenjualBarang,KondisiBarang,UkuranBarang,BrandBarang,WarnaBarang,KategoriBarang,Deskripsi,GambarBarang, "Belum Terjual");
         } catch (SQLException e) {
             // TODO Auto-generated catch block
             e.printStackTrace();